requirements.txt
-i https://pypi.org/simple
certifi==2020.4.5.1
chardet==3.0.4
idna==2.9
python-decouple==3.3
python-telegram==0.12.0
requests==2.23.0
tinydb==4.1.1
urllib3==1.25.9
